Rename package and constants
