+++ src/main/java/utility/Constant.java
@@ -3,8 +3,8 @@ package utility;
 public class Constant {
 
     // Formatted Replies
-    public static final Integer MAX_TIME_ARRAY_LENGTH = 3;
-    public static final Integer MIN_TIME_ARRAY_LENGTH = 2;
+    public static final Integer MAX_RUNTIME_ARRAY_LENGTH = 3;
+    public static final Integer MIN_RUNTIME_ARRAY_LENGTH = 2;
     public static final String PARTITION_LINE = "____________________________________________________________";
 
 }
