+++ src/main/java/workouts/Workout.java
@@ -1,4 +1,4 @@
-package exercise;
+package workouts;
 import java.time.LocalDate;
 import java.time.format.DateTimeParseException;
 import java.time.format.DateTimeFormatter;
