+++ src/test/java/workouts/ParserTest.java
@@ -1,6 +1,5 @@
-package exercise;
+package workouts;
 
-import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.AfterAll;
